[DOM8-73] fixed several bugs with matchmaking

+++ Dominion-8/src/com/tpps/technicalServices/network/login/SQLHandling/SQLStatisticsHandler.java
@@ -84,7 +84,7 @@ public class SQLStatisticsHandler {
 			stmt.setString(1, nickname);
 			ResultSet rs = stmt.executeQuery();
 			rs.next();
-			return rs.getInt("playtime");
+			return rs.getLong("playtime");
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
